Significant changes to Type Checker and Functions
Removed 'return_type' from 'Function' struct and passed it as argument to 'FuncDef' statement. The 'Return' statement now requires a 'Name' argument to help localize its function, which is necessary to evaluate the return type of a function. Additionaly, Statement checking was added to Type Checker and the Environment of Type Checker was unified with the Environment from Interpreter: they now share an Environment constituted of <Name, (EnvValue, Type)>, where 'EnvValue' for expressions are now optional. This is necessary to type check arguments passed to functions without evaluating the expressions themselves.